Harden large-scale load and performance campaigns

+++ crates/borondns-core/src/zone_image_tests/additionals_nsec_trie.rs
@@ -351,7 +351,7 @@
             &mut names,
             BlobRange {
                 offset: 0,
-                len: owner_wire.len() as u32,
+                len: owner_wire.len() as u64,
             },
             "names",
         )
